+++ backend/canisters/airdrop_bot/api/src/updates/handle_direct_message.rs
@@ -1,1 +1,1 @@
-pub use bot_api::handle_direct_message::{Response::*, *};
+pub use legacy_bot_api::handle_direct_message::{Response::*, *};